11-12二维码

+++ app/src/main/java/box/lilei/box_client/client/biz/impl/RoadBizImpl.java
@@ -50,7 +50,7 @@ public class RoadBizImpl implements RoadBiz {
             }else{
                 goods.setGoodsSaleState(Goods.SALE_STATE_OUT);
             }
-            goods.setGoodsPrice(bean.getPrice()/100);
+            goods.setGoodsPrice((double) bean.getPrice()/100);
             goodsBean = bean.getGoodsBean();
             if (goodsBean != null) {
                 goods.setGoodsId(goodsBean.getId());
